+++ app-lib/php/inline/coin-mining-staking/pos/hive-power-interest-calculator.php
