+++ app-lib/php/inline/coin-mining-staking/pow/dynamic-settings.php
