+++ app-lib/php/inline/coin-mining-staking/pow/earned-daily.php
