+++ app-lib/php/inline/coin-mining-staking/pow/profit-calculation.php
