+++ app-lib/php/inline/coin-mining-staking/pow/time-calculation.php
