+++ templates/interface/desktop/php/admin/admin-sections/webhook.php
@@ -6,6 +6,8 @@
 
 ?>
 
+	<p> Webhooks are added via the plugin system built into this app (when a specific plugin's "runtime_mode" is set to "webhook"). See <a href='https://raw.githubusercontent.com/taoteh1221/Open_Crypto_Tracker/main/DOCUMENTATION-ETC/PLUGINS-README.txt' target='_blank'>/DOCUMENTATION-ETC/PLUGINS-README.txt</a> for more information on plugin creation / development.</p>
+	
 	
 	<p> Coming Soon&trade; </p>
