+++ templates/interface/desktop/php/user/user-sections/mining.php
@@ -31,7 +31,7 @@
 			<fieldset class='subsection_fieldset'>
 				<legend class='subsection_legend'> <b><?=$pow_asset_data['name']?> Mining Calculator</b> </legend>
 		    
-				<?php require("app-lib/php/inline/calculators/mining/pow/render.php"); ?>
+				<?php require("app-lib/php/inline/coin-mining-staking/pow/render.php"); ?>
 				
 				
 			</fieldset>
@@ -43,7 +43,7 @@
 			<fieldset class='subsection_fieldset'>
 				<legend class='subsection_legend'> <b>HIVE Power Interest Rate / Power Down Weekly Payout Calculator</b> </legend>
 				
-				<?php require("app-lib/php/inline/calculators/mining/pos/hive-power-interest-calculator.php"); ?>
+				<?php require("app-lib/php/inline/coin-mining-staking/pos/hive-power-interest-calculator.php"); ?>
 				
 			</fieldset>
